Rename BoxWidget folder to BoxWidgets and with this rename the namespace to reflect the content
